rename: 기존 Auth 관련 hook 폴더 구조 수정
