+++ src/context/auth/useAuth.js
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { AuthContext } from '../context/AuthProvider';
+import { AuthContext } from './AuthContext';
 
 export const useAuth = () => {
   return useContext(AuthContext);
