+++ src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App.jsx';
-import { AuthProvider } from './context/AuthProvider.jsx';
-import { GlobalStyle } from './styles/GlobalStyle.jsx';
+import { AuthProvider } from './context/auth/AuthProvider';
+import { GlobalStyle } from './styles/GlobalStyle';
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
